Change the Numeric as integer

+++ src/monkey/lexer/lexer.go
@@ -85,7 +85,7 @@ func (l *Lexer) NextToken() token.Token {
 			tok.Type = token.LookupIdent(tok.Literal)
 			return tok
 		} else if isDigit(l.ch) {
-			tok.Type = token.NUMERIC
+			tok.Type = token.INT
 			tok.Literal = l.readNumeric()
 			return tok
 		} else {
